repeat order

+++ User/sequence/sequence.h
@@ -3,5 +3,5 @@
 #include "stm32f4xx.h"
  
 
- u8* createSequence(int initial[],int initNumber);
+ u8* createSequence(u8 initial[],u8 initNumber);
 #endif
